Phucmt at home

+++ app/src/main/java/vn/gas/thq/ui/vitri/ToaDoModel.kt
@@ -1,6 +1,6 @@
 package vn.gas.thq.ui.vitri
 
 class ToaDoModel {
-    var lat: Int? = null
-    var lng: Int? = null
+    var lat: Float? = null
+    var lng: Float? = null
 }
